+++ src/pynwb/ndx_ophys_devices/testing/__init__.py
@@ -1,6 +1,4 @@
 from ._mock import (
-    mock_DeviceModel,
-    mock_DeviceInstance,
     mock_ViralVector,
     mock_ViralVectorInjection,
     mock_Indicator,
@@ -27,8 +25,6 @@ from ._mock import (
 )
 
 __all__ = [
-    "mock_DeviceModel",
-    "mock_DeviceInstance",
     "mock_ViralVector",
     "mock_ViralVectorInjection",
     "mock_Indicator",